Rename client route components with new form abbreviations
